Split tokensource
